Conflicts:
	Firmware/sketches/Teensy/.idea/workspace.xml

+++ Firmware/sketches/Teensy/src/framework/views/View.cpp
@@ -28,7 +28,7 @@ View::~View()
 	}
 }
 
-View::View(uint16_t x, uint16_t y, uint16_t width, uint16_t height)
+View::View(int x, int y, int width, int height)
 {
 	_frame = Rect(x,y,width,height);
 	_opaque = false;
