+++ Firmware/sketches/Teensy/src/framework/views/View.h
@@ -39,7 +39,7 @@ class View: public UIElement, public AnimatableObject, Object
 #pragma mark Constructor
 public:
 	~View();
-	View(uint16_t x, uint16_t y, uint16_t width, uint16_t height);
+	View(int x, int y, int width, int height);
 	View(Rect frame);
 
 #pragma mark UIElement
